src/synapse/requirements.txt
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-storage-blob==12.20.0
azure-storage-file-datalake==12.15.0
azure-synapse==0.1.1
azure-synapse-artifacts==0.19.0
azure-synapse-spark==0.7.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
idna==3.7
isodate==0.6.1
msal==1.29.0
msal-extensions==1.2.0
msrest==0.7.1
numpy==2.0.0
oauthlib==3.2.2
pandas==2.2.2
portalocker==2.10.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
